Move most JS files also
