+++ packages/primary/tests/communityMaps.test.ts
@@ -15,10 +15,7 @@ test("there are exactly 103 official city maps", async ({ page }) => {
   );
 
   const communityMapIndex = toggleValues.indexOf("Community Maps");
-  // Consider removing if statement after first community map is implemented
-  if (communityMapIndex > -1) {
-    const officialMapIndex =
-      communityMapIndex - toggleValues.indexOf("Official Maps");
-    expect(officialMapIndex).toEqual(103);
-  }
+  const officialMapIndex =
+    communityMapIndex - toggleValues.indexOf("Official Maps");
+  expect(officialMapIndex).toEqual(103);
 });
