+++ packages/scripts/package.json
@@ -9,7 +9,6 @@
     "update-lots": "tsx src/update-lots.ts"
   },
   "dependencies": {
-    "@prn-parking-lots/primary": "workspace:*",
     "@prn-parking-lots/shared": "workspace:*",
     "ts-results": "^3.3.0"
   },
